:construction:
В консоли
rules.js:3Uncaught Error: Cannot find module "../templates/game1"(…)

при сборке
ERROR in ./templates/game1.js
Module build failed: SyntaxError: 'import' and 'export' may only appear
at the top level (70:0)
