+++ js/main.js
@@ -1,25 +1,8 @@
 import intro from '../templates/intro';
 
-import greeting from '../templates/greeting';
-import rules from '../templates/rules';
-import game1 from '../templates/game1';
-import game2 from '../templates/game2';
-import game3 from '../templates/game3';
-import stats from '../templates/stats';
+import draw from '../js/draw.js';
 
 (() => {
-
-  // Rules
-  let rulesSubmit = rules.querySelector('.rules__button');
-
-  rules.querySelector('.rules__input').oninput = (evt) => {
-    if (evt.target.value) {
-      rulesSubmit.removeAttribute('disabled');
-    } else {
-      rulesSubmit.setAttribute('disabled', '');
-    }
-  };
-
   // Slides changer
 
   let mainElement = document.getElementById('main');
@@ -31,34 +14,5 @@ import stats from '../templates/stats';
   switcher.style.cssText = 'text-align: center';
   mainElement.after(switcher);
 
-  let slides = [
-    intro,
-    greeting,
-    rules,
-    game1,
-    game2,
-    game3,
-    stats
-  ];
-  let current = -1;
-
-  let select = (index) => {
-    current = index;
-    mainElement.innerHTML = '';
-    mainElement.appendChild(slides[index]);
-  };
-
-  document.querySelector('.next').onclick = (e) => {
-    e.preventDefault();
-
-    select(current + 1);
-  };
-
-  document.querySelector('.prev').onclick = (e) => {
-    e.preventDefault();
-
-    select(current - 1);
-  };
-
-  select(0);
+  draw(intro);
 })();
