+++ templates/game1.js
@@ -1,4 +1,7 @@
-import getElementFromTemplate from '../js/getElementFromTemplate';
+import getElementFromTemplate from '../js/getElementFromTemplate.js';
+
+import game2 from '../templates/game2.js';
+import draw from '../js/draw.js';
 
 const game1 = getElementFromTemplate(`<header class="header">
     <div class="header__back">
@@ -55,5 +58,13 @@ const game1 = getElementFromTemplate(`<header class="header">
       </ul>
     </div>
   </div>`);
+
+const answers = game1.querySelectorAll('.game__answer');
+
+for (const answer of answers) {
+  answer.onclick = (e) => {
+    e.preventDefault();
+    draw(game2);
+};
 
 export default game1;
