+++ templates/game3.js
@@ -1,5 +1,8 @@
 import getElementFromTemplate from '../js/getElementFromTemplate';
 
+import stats from '../templates/stats';
+import draw from '../js/draw.js';
+
 const game3 = getElementFromTemplate(`<header class="header">
     <div class="header__back">
       <span class="back">
@@ -43,5 +46,14 @@ const game3 = getElementFromTemplate(`<header class="header">
     </div>
   </div>
 `);
+
+const answers = game3.querySelectorAll('.game__option');
+
+for (const answer of answers) {
+  answer.onclick = (e) => {
+    e.preventDefault();
+    draw(stats);
+  };
+}
 
 export default game3;
